+++ async_upnp_client/discovery.py
@@ -92,7 +92,7 @@ async def async_discover(timeout: int = SSDP_MX,
     # create socket
     sock = socket.socket(socket.AF_INET, socket.SOCK_DGRAM, socket.IPPROTO_UDP)
     sock.setsockopt(socket.IPPROTO_IP, socket.IP_MULTICAST_TTL, timeout)
-    sock.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
+    # sock.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
     if source_ip:
         sock.bind((source_ip, 0))