+++ async_upnp_client/igd.py
@@ -52,6 +52,11 @@ class IgdDevice(UpnpProfileDevice):
 
     # pylint: disable=too-many-public-methods
 
+    DEVICE_TYPES = [
+        'urn:schemas-upnp-org:device:InternetGatewayDevice:1',
+        'urn:schemas-upnp-org:device:InternetGatewayDevice:2',
+    ]
+
     _SERVICE_TYPES = {
         'WANPPPC': {
             'urn:schemas-upnp-org:service:WANPPPConnection:1',