+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changes
 0.13.3 (unreleased)
 
 - Add option to `upnp-client` to set timeout for device communication/discovery
+- Add option to be strict (default false) with regard to invalid data
 
 
 0.13.2 (2018-11-11)