+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changes
 
 - Add option to `upnp-client` to set timeout for device communication/discovery
 - Add option to be strict (default false) with regard to invalid data
+- Add more error handling to `upnp-client`
 
 
 0.13.2 (2018-11-11)